+++ pmd-dist/src/test/java/net/sourceforge/pmd/it/AbstractBinaryDistributionTest.java
@@ -5,13 +5,11 @@
 package net.sourceforge.pmd.it;
 
 import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 
-import org.apache.commons.io.FileUtils;
-import org.junit.AfterClass;
 import org.junit.BeforeClass;
+import org.junit.ClassRule;
+import org.junit.rules.TemporaryFolder;
 
 import net.sourceforge.pmd.PMDVersion;
 
@@ -21,6 +19,9 @@ public abstract class AbstractBinaryDistributionTest {
         return new File(".", "target/pmd-bin-" + PMDVersion.VERSION + ".zip");
     }
 
+    @ClassRule
+    public static TemporaryFolder folder = new TemporaryFolder();
+
     /**
      * The temporary directory, to which the binary distribution will be extracted.
      * It will be deleted again after the test.
@@ -29,16 +30,9 @@ public abstract class AbstractBinaryDistributionTest {
 
     @BeforeClass
     public static void setupTempDirectory() throws Exception {
-        tempDir = Files.createTempDirectory("pmd-it-test-");
+        tempDir = folder.newFolder().toPath();
         if (getBinaryDistribution().exists()) {
             ZipFileExtractor.extractZipFile(getBinaryDistribution().toPath(), tempDir);
-        }
-    }
-
-    @AfterClass
-    public static void cleanupTempDirectory() throws IOException {
-        if (tempDir != null && tempDir.toFile().exists()) {
-            FileUtils.forceDelete(tempDir.toFile());
         }
     }
 }